Initial commit
This commit has an Astro plugin that is building and 90% working. It can
take an array of stylesheets with `@font-face` declarations, extract
them, download the associated fonts into the app's `public` directory,
and then have them be processed by fontaine to generate fallback
`@font-face` declarations that will limit content shifting.

Sadly, this isn't fully working with production builds nor does it
inline these `@font-face` declarations into `<style>` tag such that
they immediately work. This is a limitation of using the drop-in Vite
plugin. In order to get around this, I need to directly use fontaine's
primitives, like the nuxt version of it does.

I would also LOVE to be able to define which base fallback fonts are
used per font. In my test site, it will use sans-serif fonts for
everything while loading, which looks terrible.